test(tvc): consolidate secrets CLI surface tests into one file
